Safari-611.1.10/LayoutTests/http/tests/media/media-document-expected.txt


Tests that a media document can open a media URL with an ambiguous extension.

EXPECTED (frame.contentDocument.getElementsByTagName('video')[0].error == 'null') OK
END OF TEST
